% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_cdist.R
\name{cdist}
\alias{cdist}
\title{Calcutating the mean distance among the CpG sites contained in the analysed region}
\usage{
cdist(matrix)
}
\arguments{
\item{matrix}{Epiallele binary matrix for one genomic region}
}
\value{
Integer indicating the mean distance between CpG in the interval
}
\description{
Calcutating the mean distance among the CpG sites contained in the analysed region
}
