% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23_diffAdjust.R
\name{diffAdjust}
\alias{diffAdjust}
\title{diffAdjust}
\usage{
diffAdjust(
  data,
  metadata,
  samples,
  statistic,
  groupcol,
  covariate,
  cores = 1,
  reduce = FALSE
)
}
\arguments{
\item{metadata}{your samples metadata. The input file must contain the columns named "Group" and "Samples"}

\item{intervals_list}{list of samples intervals. It corresponds to the output 'epi' from the epiAnalysis function}
}
\value{
Dataframe with statistics
}
\description{
diffAdjust
}
