% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23_diffModel.R
\name{diffModel}
\alias{diffModel}
\title{Getting regions which differ for one statistic (such as Shannon Entropy, etc...) through linear model and possibly adjusting for a covariate}
\usage{
diffModel(
  intervals_list,
  metadata,
  samples,
  statistic,
  groupcol,
  covariate,
  min.per.group = 2,
  min.groups = 2,
  cores = 1,
  reduce = FALSE
)
}
\arguments{
\item{intervals_list}{A list object. It corresponds to a list containing the summary outputs obtained through epiAnalysis() function (the ones containing the summary statistics, such as Shannon Entropy, Mean CpGs distance, etc...).}

\item{metadata}{Samples metadata provided as table. This input data should contain specific columns for samples names and groups distinction.}

\item{statistic}{Character indicating the column name that contains the statistic vector the user wants to use to perform the test (e.g., "Shannon", "mean_met").}

\item{groupcol}{Character indicating the column name of the Groups in the metadata used to compare the statistics.}

\item{covariate}{Character indicating the column name of the covariate to be used to adjust the linear model.}

\item{min.per.group}{An integer indicating the minimum number of samples for group required to perform the test.}

\item{min.groups}{An integer indicating the minimum number of groups wanted to perform the test.}

\item{cores}{An integer indicating the number of cores to be used to perform the computation.}

\item{reduce}{Logical indicating whether adjacent overlapping intervals should be reduced as a unique interval or not.}
}
\value{
A dataframe containing the statistical test results.
}
\description{
This function is implemented to get different significant regions using a linear model adjusted for a covariate (different Points in time, different tissue, etc...).
}
\examples{
data(epistats)
samples_list <- list(Sample1_intervals.bed,
                     Sample2_intervals.bed,
                     Sample3_intervals.bed,
                     Sample4_intervals.bed)

diffresult <- diffAdjust(intervals_list = samples_list,
                         metadata = ann,
                         samples = "Samples",
                         statistic = "Shannon",
                         groupcol = "Group",
                         covariate = "Time",
                         cores = 40,
                         reduce = FALSE)
}
