% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_epi.R
\name{epi}
\alias{epi}
\title{Calculating the number of different epialleles observed for one genomic region}
\usage{
epi(matrix)
}
\arguments{
\item{matrix}{Epiallele binary matrix.}
}
\value{
A numeric indicating the number of observed epialleles in one interval.
}
\description{
Calculating the number of different epialleles observed for one genomic region
}
