% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_epiAnalysis.R
\name{epiAnalysis}
\alias{epiAnalysis}
\title{epiAnalysis}
\usage{
epiAnalysis(
  align,
  bin,
  bisu.Thresh = 0.8,
  stranded = FALSE,
  mode = "CG",
  remove.Amb = TRUE,
  genome,
  pathDir,
  cores = 10,
  retain.reads = FALSE,
  get.cPos = FALSE,
  myfuns
)
}
\arguments{
\item{align}{GAlignments object}

\item{bin}{intervals from makeWindows or makeEpialleles}

\item{bisu.Thresh}{integer to set bisu efficiency threshold}

\item{stranded}{logical}

\item{mode}{pattern character}

\item{remove.Amb}{logical}

\item{genome}{fasta}

\item{pathDir}{directory to save output}

\item{cores}{num of cores to use}

\item{retain.reads}{logical}

\item{get.cPos}{logical}

\item{myfuns}{list of functions}
}
\value{
list of dataframes
}
\description{
epiAnalysis
}
