% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_epidiv.R
\name{epidiv}
\alias{epidiv}
\title{epidiv}
\usage{
epidiv(samples, region, metadata)
}
\arguments{
\item{samples}{list of epimatrices}

\item{region}{one region of interest}

\item{metadata}{samples metadata}
}
\value{
list
}
\description{
epidiv
}
