% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_epiplot.R
\name{epiplot}
\alias{epiplot}
\title{epiplot}
\usage{
epiplot(samples, region, metadata)
}
\arguments{
\item{samples}{list of epimatrices}

\item{region}{one region of interest}

\item{metadata}{samples metadata}
}
\value{
list
}
\description{
epiplot
}
