% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_epiStat.R
\name{epiStat}
\alias{epiStat}
\title{epiStat}
\usage{
epiStat(
  sample_list,
  metadata,
  colgroups,
  colsamples,
  rmUnmeth = FALSE,
  cores = 1,
  minGroups = 2,
  minSampleSize = 2,
  reduce = FALSE
)
}
\arguments{
\item{sample_list}{list of epiMatrix coming from epiAnalysis function}

\item{metadata}{dataframe. Your samples metadata}

\item{colgroups}{character indicating the colname of Groups vector}

\item{colsamples}{character indicating the colname of Samples vector}

\item{rmUnmeth}{logical indicating if 0 methylated species should be removed of not from the analysis}

\item{cores}{num of cores}

\item{minGroups}{integer indicating the minimum number of unique Groups should be used for the dissimilarity analysis among samples}

\item{minSampleSize}{integer indicating the minimum number of samples per group needed to perform the statistical analysis}

\item{reduce}{logical indicating if intervals with the same statistics should be reduced or not}
}
\value{
dataframe with rows indicating the genomic regions analysed and columns containing the statistics
}
\description{
epiStat
}
