% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_filterBAM.R
\name{filterBAM}
\alias{filterBAM}
\title{filterBAM}
\usage{
filterBAM(algn, keepStChr = TRUE, keepM = FALSE, retainChr = NULL)
}
\arguments{
\item{algn}{GAlignments object}

\item{keepStChr}{logical}

\item{keepM}{logical}

\item{retainChr}{logical}
}
\value{
filtered GAlignments object
}
\description{
filterBAM
}
