% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_filterByCoverage.R
\name{filterByCoverage}
\alias{filterByCoverage}
\title{filterByCoverage}
\usage{
filterByCoverage(algn, threshold, minsize)
}
\arguments{
\item{algn}{GAlignments object}

\item{threshold}{integer}

\item{minsize}{integer}
}
\value{
filtered GRanges object
}
\description{
filterByCoverage
}
