% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_genCoveredRegONT.R
\name{genCoveredRegONT}
\alias{genCoveredRegONT}
\title{Selecting regions with a minimum user-defined coverage}
\usage{
genCoveredRegONT(sparseMat, threshold = 10)
}
\arguments{
\item{sparseMat}{Sparse matrix generated with loadInputONT().}

\item{threshold}{Integer indicating the minimum coverage required to keep genomic regions.}
}
\value{
filtered GRanges object.
}
\description{
The functions allows the selection of genomic regions that satisfy a minimum coverage threshold defined by the user as a parameter.
}
