% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_loadInput.R
\name{loadInput}
\alias{loadInput}
\title{loadInput}
\usage{
loadInput(bamfile, genomefile)
}
\arguments{
\item{bamfile}{path to bamfile}

\item{genomefile}{path to genome/fasta}
}
\value{
data
}
\description{
loadInput
}
