% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_makeBins.R
\name{makeBins}
\alias{makeBins}
\title{Designing target regions containing a fixed number of CpG sites}
\usage{
makeBins(
  gr,
  Genome,
  mode = "CG",
  n = 4,
  min.binsize = 1,
  max.binsize = 200,
  cores = 1
)
}
\arguments{
\item{gr}{A GRanges object obtained from the previous filtering functions.}

\item{Genome}{A DNAStringSet containing the fasta genome.}

\item{mode}{A character indicanting the pattern to be used to perform the epiallele analysis ("CG", "CA", "CT", "CC")}

\item{n}{An integer indicating the number of CpGs sites to be contained in the target region.}

\item{min.binsize}{An integer indicating the minimum length size of each bin}

\item{max.binsize}{An integer indicating the maximum length size of each bin}

\item{cores}{An integer indicating the number of cores to be used for the process}
}
\value{
A GRanges object containing the newly designed target regions with the features specified through the function parameters.
}
\description{
This functions represents one of the two methods available in the epistats package to design customised regions. makeBins() is provided to design target regions which contain a given number of CpG sites (or any of the other C* patterns).
}
\examples{
#' data <- loadInput(bamfile, genomefile)
algn <- data[[1]]

## Keeping only standard chromosomes
filtered <- filterBAM(algn,
                      keepStChr =TRUE,
                      keepM = FALSE,
                      retainChr = NULL)

## Selecting regions with a minimum coverage
covered <- filterByCoverage(algn, threshold = 50, minsize = 8)

## Designing regions containing 4 CpG sites
  regions = makeBins(gr = covered,
                     Genome = Genome,
                     mode = "CG",
                     n = 4,
                     min.binsize = 8,
                     max.binsize = 100,
                     cores = 40)
}
