% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_maxfreq.R
\name{maxfreq}
\alias{maxfreq}
\title{Reporting the epiallele observed at the highest frequency in one genomic interval}
\usage{
maxfreq(matrix)
}
\arguments{
\item{matrix}{Epialleles binary matrix.}
}
\value{
A string reporting the epiallele observed at the highest frequency in the genomic interval.
}
\description{
Reporting the epiallele observed at the highest frequency in one genomic interval
}
