% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_meanMeth.R
\name{meanMeth}
\alias{meanMeth}
\title{Calculating the average DNA methylation for one interval}
\usage{
meanMeth(matrix)
}
\arguments{
\item{matrix}{Epialleles binary matrix.}
}
\value{
A numeric indicating the average DNA methylation of one genomic region
}
\description{
Calculating the average DNA methylation for one interval
}
