% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_meanMeth.R
\name{meanMethONT}
\alias{meanMethONT}
\title{Calculating the average DNA methylation for one interval for ONT data}
\usage{
meanMethONT(matrix, methylated_code = 1L)
}
\arguments{
\item{matrix}{Epialleles matrix.}

\item{methylated_code}{A numeric value indicating the methylation status in the epialelles matrix.}
}
\value{
A numeric indicating the average DNA methylation of one genomic region
}
\description{
Calculating the average DNA methylation for one interval for ONT data
}
