% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_numReads.R
\name{num_reads}
\alias{num_reads}
\title{Counting the number of reads mapping on one genomic interval.}
\usage{
num_reads(matrix)
}
\arguments{
\item{matrix}{Epialleles binary matrix.}
}
\value{
An integer indicating the coverage observed in the genomic region.
}
\description{
Counting the number of reads mapping on one genomic interval.
}
