% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_numCG.R
\name{numCG}
\alias{numCG}
\title{Counting the number of CpG sites contained in one interval}
\usage{
numCG(matrix)
}
\arguments{
\item{matrix}{Epialleles binary matrix.}
}
\value{
An integer indicating the number of CpGs displayed for one genomic interval.
}
\description{
Counting the number of CpG sites contained in one interval
}
