% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_oneStat.R
\name{oneStat}
\alias{oneStat}
\title{oneStat}
\usage{
oneStat(sample_list, region, metadata)
}
\arguments{
\item{sample_list}{list of samples epiMatrices coming from epiAnalysis function}

\item{region}{genomic coordinate of your region of interest}

\item{metadata}{dataframe. Your samples metadata}
}
\value{
Dataframe with statistics
}
\description{
oneStat
}
