% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24_regsToPathway.R
\name{regsToPathway}
\alias{regsToPathway}
\title{regsToPathways}
\usage{
regsToPathway(stat_out, genome)
}
\arguments{
\item{stat_out}{list of samples intervals. It corresponds to the output 'epi' from the epiAnalysis function}

\item{genome}{your samples metadata. The input file must contain the columns named "Group" and "Samples"}
}
\value{
list of Dataframes
}
\description{
regsToPathways
}
