% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_runCCA.R
\name{runCCA}
\alias{runCCA}
\title{runCCA}
\usage{
runCCA(samples, region, metadata, printData = FALSE)
}
\arguments{
\item{samples}{list of epimatrices}

\item{region}{one region of interest}

\item{metadata}{samples metadata}
}
\value{
list
}
\description{
runCCA
}
