% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_runRDA.R
\name{runRDA}
\alias{runRDA}
\title{Running principal component analysis for one genomic interval}
\usage{
runRDA(samples, region, printData = FALSE, rmUnmeth = FALSE)
}
\arguments{
\item{samples}{A list object containing the epiallele composition matrices from all the samples of the dataset.}

\item{region}{A character indicating the regionID wanted to perform the analysis.}

\item{printData}{Logical indicating whether the epiallele matrix should be printed in the standard output or not (Default = FALSE).}

\item{rmUnmeth}{Logical indicating if 0 methylated epialleles should be removed or not from the analysis.}
}
\value{
Principal Component analysis results and plot.
}
\description{
Running principal component analysis for one genomic interval
}
