% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_shannon.R
\name{shannon}
\alias{shannon}
\title{Calcuting the Shannon Entropy in one genomic region.}
\usage{
shannon(matrix)
}
\arguments{
\item{matrix}{Epialleles binary matrix.}
}
\value{
A numeric indicating Shannon Entropy value relative to one genomic interval.
}
\description{
Calcuting the Shannon Entropy in one genomic region.
}
