% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_singleton.R
\name{singleton}
\alias{singleton}
\title{Calculating the number of epialleles with just one observation}
\usage{
singleton(matrix)
}
\arguments{
\item{matrix}{Epiallele binary matrix}
}
\value{
An integer indicating the number of epiallele with a unique observation in a given interval
}
\description{
Calculating the number of epialleles with just one observation
}
