% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23_test2Shannon.R
\name{test2Shannon}
\alias{test2Shannon}
\title{test2Shannon}
\usage{
test2Shannon(intervals_list, metadata)
}
\arguments{
\item{intervals_list}{list of samples intervals. It corresponds to the output 'epi' from the epiAnalysis function}

\item{metadata}{your samples metadata. The input file must contain the columns named "Group" and "Samples"}
}
\value{
Dataframe with statistics
}
\description{
test2Shannon
}
