% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22_testShannon.R
\name{testShannon}
\alias{testShannon}
\title{testShannon}
\usage{
testShannon(intervals_list, metadata)
}
\arguments{
\item{intervals_list}{list of samples intervals. It corresponds to the output 'epi' from the epiAnalysis function}

\item{metadata}{your samples metadata. The input file must contain the columns named "Group" and "Samples"}
}
\value{
Dataframe with statistics and the test used to compare the Shannon between samples
}
\description{
testShannon
}
