% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22_diffStat.R
\name{diffStat}
\alias{diffStat}
\title{diffStat}
\usage{
diffStat(
  intervals_list,
  metadata,
  colsamples,
  statistic,
  groupcol,
  cores = 1,
  reduce = FALSE
)
}
\arguments{
\item{intervals_list}{list of samples intervals. It corresponds to the output 'epi' from the epiAnalysis function}

\item{metadata}{your samples metadata. The input file must contain the columns named "Group" and "Samples"}

\item{statistic}{parameter indicating the colname of the statistic the user wants to use to perform the test}

\item{groupcol}{character indicating the Group colname in the metadata used to compare the statistics}

\item{cores}{number of cores to use}

\item{reduce}{logical indicating whether adiacent overlapping intervals should be reduced or not}
}
\value{
Table containing the p-values obtained comparing the given statistic between samples of different groups
}
\description{
diffStat
}
