% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_epiAnalysis.R
\name{epiAnalysis}
\alias{epiAnalysis}
\title{epiAnalysis}
\usage{
epiAnalysis(
  align,
  bin,
  aname,
  threshold,
  pathDir,
  genome,
  myfuns,
  bisu.Thresh = 0,
  stranded = FALSE,
  mode = "CG",
  remove.Amb = TRUE,
  cores = 1,
  retain.reads = FALSE,
  get.cPos = FALSE
)
}
\arguments{
\item{align}{GAlignments object}

\item{bin}{intervals from makeWindows or makeEpialleles}

\item{aname}{character indicating the analysis name}

\item{threshold}{integer indicating the minimum coverage required to analyse an interval}

\item{pathDir}{directory to save output}

\item{genome}{fasta}

\item{myfuns}{list of functions}

\item{bisu.Thresh}{integer to set bisu efficiency threshold}

\item{stranded}{logical}

\item{mode}{pattern character}

\item{remove.Amb}{logical}

\item{cores}{num of cores to use}

\item{retain.reads}{logical}

\item{get.cPos}{logical}
}
\value{
list of dataframes
}
\description{
epiAnalysis
}
