% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_epidiv.R
\name{epidiv}
\alias{epidiv}
\title{Finding the epiallele species which are contributing most to the dissimilarity among groups}
\usage{
epidiv(samples, region, metadata)
}
\arguments{
\item{samples}{A list object containing the epiallele composition matrices from all the samples of the dataset.}

\item{region}{A string containing the regionID wanted to perform the analysis.}

\item{metadata}{A dataframe object containing samples metadata. Dataframe should contain dedicated columns for samples IDs and the one indicating the group they belong to.}
}
\value{
A list containing the dissimilarity results.
}
\description{
The function returns the information regarding which epialleles are the ones contributing for the dissimilarity of samples among groups.
}
