% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_epiStat.R
\name{epiStat}
\alias{epiStat}
\title{epiStat}
\usage{
epiStat(sample_list, metadata, cores)
}
\arguments{
\item{sample_list}{list of epiMatrix coming from epiAnalysis}

\item{metadata}{samples metadata}

\item{cores}{num of cores}
}
\value{
dataframe w/ stats
}
\description{
epiStat
}
