% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_filterBAM.R
\name{filterBAM}
\alias{filterBAM}
\title{Filtering function}
\usage{
filterBAM(algn, keepStChr = TRUE, keepM = FALSE, retainChr = NULL)
}
\arguments{
\item{algn}{GAlignments object containing aligned reads}

\item{keepStChr}{Logical indicating whether only standard chromosomes should be kept}

\item{keepM}{Logical indicating whether the mitochondrial chromosome should be kept for the analysis or not}

\item{retainChr}{Vector containing the chromosomes to be analysed as characters. Note that the strings should correspond to the seqlevels contained in the GAlignment object.}
}
\value{
A filtered GAlignments object.
}
\description{
This function is provided to filter only selected regions from bamfiles (specific chromosomes of interest, etc...).
}
