% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_getEpimatrix.R
\name{getEpimatrix}
\alias{getEpimatrix}
\title{getEpimatrix}
\usage{
getEpimatrix(samples, region)
}
\arguments{
\item{samples}{list of samples epiMatrices coming from epiAnalysis function}

\item{region}{genomic coordinate of your region of interest}
}
\value{
Epiallele community data matrix
}
\description{
getEpimatrix
}
