% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19_getEpimatrix.R
\name{getEpimatrix}
\alias{getEpimatrix}
\title{Getting the epialleles matrix for one genomic region}
\usage{
getEpimatrix(samples, region)
}
\arguments{
\item{samples}{A list object containing the epiallele composition matrices from all the samples of the dataset.}

\item{region}{A character indicating the regionID wanted to perform the analysis.}
}
\value{
Epiallele community data matrix.
}
\description{
The function gives as output the epiallele composition matrix relative to one genomic interval.
}
