% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_loadInput.R
\name{loadInput}
\alias{loadInput}
\title{Loading Input data.}
\usage{
loadInput(bamfile, genomefile)
}
\arguments{
\item{bamfile}{path to bamfile to be analysed. Note that folder containing the bamfile should also contain the bam index.}

\item{genomefile}{path to genome (FASTA format).}
}
\value{
A list containg a GAlignment object and a DNAStringSet object.
}
\description{
Loading Input data.
}
