% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_makeBins.R
\name{makeBins}
\alias{makeBins}
\title{makeBins}
\usage{
makeBins(
  gr,
  Genome,
  mode = "CG",
  n = 4,
  min.binsize = 1,
  max.binsize = 200,
  cores = 1
)
}
\arguments{
\item{gr}{GRanges}

\item{Genome}{fasta file}

\item{mode}{character pattern}

\item{n}{num CpG in the bin}

\item{min.binsize}{min length of a bin}

\item{max.binsize}{max length of a bin}

\item{cores}{number of cores to use}
}
\value{
GRanges
}
\description{
makeBins
}
