% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_makeWindows.R
\name{makeWindows}
\alias{makeWindows}
\title{makeWindows}
\usage{
makeWindows(gr, window, step, Genome, mode, min.C, max.C, cores)
}
\arguments{
\item{gr}{GRanges}

\item{window}{length size of the sliding window}

\item{step}{step size to slide}

\item{Genome}{fasta Genome}

\item{mode}{pattern character}

\item{min.C}{min num of CpG in the sliding window}

\item{max.C}{max num of CpG in the sliding window}

\item{cores}{num of cores to use}
}
\value{
GRanges
}
\description{
makeWindows
}
