% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_meanMeth.R
\name{meanMeth}
\alias{meanMeth}
\title{meanMeth}
\usage{
meanMeth(matrix)
}
\arguments{
\item{matrix}{epialleles matrix}
}
\value{
numeric indicating the mean meth of the entire region
}
\description{
meanMeth
}
