% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_meanMeth.R
\name{meanMethONT}
\alias{meanMethONT}
\title{Calculating the average DNA methylation for a not binary epialleles matrix.}
\usage{
meanMethONT(matrix, param = list(meth_code = 1L, unmeth_code = 0L))
}
\arguments{
\item{matrix}{Epialleles matrix.}

\item{param}{A list of two numeric values indicating the codes for methylated and unmethylated status in the epialleles matrix.}
}
\value{
A numeric indicating the average DNA methylation of one genomic region
}
\description{
Calculating the average DNA methylation for a not binary epialleles matrix.
}
