% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_numReads.R
\name{num_reads}
\alias{num_reads}
\title{numReads}
\usage{
num_reads(matrix)
}
\arguments{
\item{matrix}{epialleles matrix}
}
\value{
integer indicating the depth of a given region
}
\description{
numReads
}
