% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_oneTest.R
\name{oneTest}
\alias{oneTest}
\title{oneTest}
\usage{
oneTest(
  sample_list,
  region,
  metadata,
  rmUnmeth = FALSE,
  minGroups = 2,
  minSampleSize = 2,
  printData = FALSE
)
}
\arguments{
\item{sample_list}{list of samples epiMatrices coming from epiAnalysis function}

\item{region}{genomic coordinate of your region of interest}

\item{metadata}{dataframe. Your samples metadata}
}
\value{
Dataframe with statistics
}
\description{
oneTest
}
