% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_pairtest.R
\name{pairtest}
\alias{pairtest}
\title{pairtest}
\usage{
pairtest(samples, region, metadata)
}
\arguments{
\item{samples}{list of epimatrices}

\item{region}{one region of interest}

\item{metadata}{samples metadata}
}
\value{
dataframe
}
\description{
pairtest
}
