% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_pairtest.R
\name{pairtest}
\alias{pairtest}
\title{Post-hoc analysis to find which groups are most contributing to the dissimilatity.}
\usage{
pairtest(samples, region, metadata)
}
\arguments{
\item{samples}{A list object containing the epiallele composition matrices from all the samples of the dataset.}

\item{region}{A string containing the regionID wanted to perform the analysis.}

\item{metadata}{A dataframe object containing samples metadata. Dataframe should contain dedicated columns for samples IDs and the one indicating the group they belong to.}
}
\value{
A dataframe containing the p-values for each group comparison, when groups are more than two.
}
\description{
This function perform a pairwise PERMANOVA analysis in order to figure out which groups are differing between each other.
}
