% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_runRDA.R
\name{runRDA}
\alias{runRDA}
\title{runRDA}
\usage{
runRDA(samples, region, printData = FALSE)
}
\arguments{
\item{samples}{list of epimatrices}

\item{region}{one region of interest}
}
\value{
PCA plot
}
\description{
runRDA
}
