% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_singleton.R
\name{singleton}
\alias{singleton}
\title{singleton}
\usage{
singleton(matrix)
}
\arguments{
\item{matrix}{epialleles matrix}
}
\value{
integer indicating how many species are represented alone
}
\description{
singleton
}
